Updated gitconfig: Moved autosign into alias
Autosigning is good and all... but having to type in the GPG password
everytime I have to make a commit has become a pain. Especially with my
current mode of work where I have intermediary commits which do not
matter to the code anyway because they get squashed before they get into
master.

What's better for me is to have this in an alias.

```
[alias]
  csm = commit -S --message
```

This way, I can work without having to sign everytime. Then when I want
to put my stamp on it, I have an alias to help me out with that.