updated vimrc: massive rehaul
My vim was getting slow, so I had to get rid of the plugins I no longer
need. The slowness may be a function of the size of the repo I am
working on now; but at least I am able to slim Vim a bit.

I also tried to make this thing neater. There is nothing better than
when all configurations are ordered properly.

The biggest change here though is to Vim-Test. This little giant got a
sister package Vim-Dispatch which uses the Vim 8 functions to delegate
commands to the shell. Vim continues uninterrupted. This... if I am to
be honest... feels like the thing which makes my Vim slower; but let's
see.