updated vimrc
Elixir editing always brought up some 1 to 4 beam.smp processes. I had
thought this was because of vim-elixir or alchemist. Turns out this was
just ALE trying to use a language server for linting. Specifying the
linters to use which did not depend on a language server solved this for
me.