updated vimrc
for some reason, when I use the super-completion tab (or something like
that, can't remember really what it's called), there are traces of
lighter grey on my screen. This is usually from the floating window
introduced by the autocomplete. To get this off, I usually would re-open
the buffer... until today.

I found (AND MAPPED!) redraw 😎