updated vimrc
oftentimes someone would ask me to Slack them a bit of code I am looking
at in my editor. If this is within a line, I can use the normal iTerm
highlight and copy. If this is a block, the usual highlight and copy
becomes a problem because the copied text comes with some line numbers.
To work around this, I have to yank stuff into the `+` register. The
movement is unnatural for me at this point, so I mapped it to the leader