added git-footprints
Have you ever wanted to know how much effect a contributor has had on a
codebase? This is what you want to use in checking that. I know this
denigrates to Management 1.0; counting lines of code and all, but I just
thought it was an interesting metric to keep. — says a lot about me (:

This version would have to be fed the output from some git magic;
precisely: `git log --authorname="" --pretty=tformat: --numstat.

The end goal is to make sure someone can type `git footprint-for author`
and get their footprints. ...I cannot confirm nor deny this isn't on a
path to rank developers based on their LOC contributions.