updated vimrc: thesaurus + projectionist
I just found out that there is a way to add a thesaurus to your Vim.
This is amazing for me because I would begin doing all my writing here.
Evernote would always be a good place to have my notes, but how cool
would it be to have my notes, poems, and stories on Github as text!!

Also there is projectionist. I know this is a good plugin, but I haven't
figured out a way to understand it yet. This is a violation of OO's ISP
but... who cares? ...fow now.