updated vimrc: leverage unimpaired
I had originally used the same keystrokes defined in unimpaired to jump
between buffers. I realized this was a tautology of sorts and decided to
remove my definition. This is because unimpaired would have more
definitions which match the motions I am used to, and defined for
myself. But it would also extend this into areas I currently do not know