updated vimrc: autowrite
Removed the autowrite per command since this is somewhat of a tautology.
It's already handled by the vim-auto-save plugin. Now, I know I should
be slimming Vim as much as possible, but hey... if someone wants to do
the honors of making sure auto-save is handled, why should I stress?