updated vimrc: per-project configurations
Experience has taught me that there is no one-size-fits-all
configuration which works for Vim. There are some projects which need
special attention. It may be because there are too many large files, or
the tools which are used in the projects are different. This means new
configuration for the editor.

This is the first case. A proof of concept. Something to show that it's
possible.

For the usecase, I found that there is some slow-down when the syntax is
on for large files. Some previous commits ago - if not the very previous
commit - I made a change which affects files with more than 1000 lines.
This takes that one step further. Instead of punish all files which
could be 1000+ lines for every project, only do it for the project you
have noticed it to be a problem in.

Also, this could be extended to setting the proper testing tools, and
mapping them to the keystrokes I am conversant in.