updated bootstrap: load pkgs into bash_profile
I noticed that the packages I loaded into bashrc weren't loading up in
this new Catalina. I think this may have something to do with the Bash
-> Zsh switch in the terminals. It prefers packages linked to in the
profile.

I read once too that there is a disparity between bashrc and
bash_profile: either something to do with TTY terminal disparity, or
something to do with shells which [do not] require logins. This could
also be the case here, but I am not certain.

It works in bash_profile; we're using bash_profile henceforth.