telescope: prefer telescope to FZF (#13)
This is a hude one. Telescope is like FZF and FZF.vim in one. It also
comes with the FZF algorithm as telescope-fzf-native.vim which makes
searches super fast. I also preferred the_silver_searcher over ripgrep,
which also increases the speed of searches.

Another addition (unnecessary but cool) is the telescope-github.vim
plugin which integrates well with GitHub's new "gh" CLI tool. There is
some delay when viewing issues because of the network call. But it
generally would come in handy to stop context-switching into a web
browser. There are mappings to switch into a web browser though.