svelte: use infinite textwidth
It was getting annoying when the indent gets reapplied for every
keystroke. This broke tags longer than 80 lines. Best I could triage,
this behaviour comes from runtime/xmlformat.vim which sets a precursor
for the html format which is heavily depended on in Svelte.

Setting an infinite textwidth allows me to trigger indentation whenever
I want to, and not be at the mercy of an over eager editor.