vim: remove vim-eunuch
This plugin did me well. I will miss `:Delete!`. But I;m sure i can
replace that with functionalities from NERDTree.

I'm removing this because it's interferring with how I write Ruby. For
some reason, the EunuchNewLine function brings the cursor to the top of
class block when I close a method block. This happens when I am defining
the class afresh. It's been going on for a while. I today to snipe it.

Plenty love for tpope. Keep doing the lords work.